+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [0-based versioning](https://0ver.org/).
 - Rework event library
 - Simplify ticker tasks in all examples
 - Prune critical section use from onesize memory allocator
+- Renamed async to coro
 
 ### Added